WIP - OIDC - Refine tests
