+++ package.json
@@ -19,8 +19,7 @@
     "lint-fix": "yarn lint --fix",
     "release": "yarn clean && yarn build && standard-version && git push --follow-tags && npm publish",
     "start": "nuxt-ts start demo",
-    "test": "yarn lint && yarn build && nuxt build test/fixture && jest",
-    "test:oidc": "yarn lint && yarn build && nuxt build test/fixture && jest test/oidc.test.js"
+    "test": "yarn lint && yarn build && nuxt build test/fixture && jest"
   },
   "dependencies": {
     "@nuxtjs/axios": "^5.12.0",
